Fix portability problems
utmp interface is a very system dependet part in Unix. There are
three different interfaces: SystemV, POSIX and BSD, and they are
incompatibles between them. Utmp was using POSIX interface, but
not all the systems implement it (for examle OpenBSD doesn't it),
so it is desirable add the code for the three interfaces. This
patch adds this new code without using a ifdef hell.